requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
numpy==2.1.3
sentence-transformers==3.2.1
